% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saffir}
\alias{saffir}
\title{saffir}
\usage{
saffir()
}
\value{
Dataframe of Saffir Simpson Hurricane Scale.
}
\description{
Returns dataframe of Saffir Simpson Hurricane Scale.
}
\details{
Classification of storms are based on wind speed. The 
\href{http://www.nhc.noaa.gov/aboutsshws.php}{Saffir-Simpson Hurricane Scale} 
is technically defined for classifying hurricanes but this will classify 
tropical storms and depressions as well.

Subtropical or Extratropical storms should not use this measure but rather 
use the status contained in the header of the text product. This function 
should be used for summaries only.
}

